+++ src/lib/Api.php
@@ -102,9 +102,29 @@ class Api
      */
     public function createSubscribers(array $data): ResponseInterface
     {
-        return $this->client->post('api/contact.php', [
-            RequestOptions::JSON => $data,
-        ]);
+        return $this->client->post(
+            'api/contact.php',
+            [
+                RequestOptions::JSON => $data,
+            ]
+        );
+    }
+
+    /**
+     * Update subscriber information.
+     *
+     * @param array $data
+     *
+     * @return ResponseInterface
+     */
+    public function updateSubscriber(array $data): ResponseInterface
+    {
+        return $this->client->post(
+            'api/contact.php',
+            [
+                RequestOptions::JSON => $data,
+            ]
+        );
     }
 
     /**